signals_bot/requirements.txt
python-telegram-bot[job-queue]==21.5
asyncpg==0.29.0
redis[asyncio]==5.0.8
loguru==0.7.2
python-dotenv==1.0.1
